+++ src/components/FontAwesomeIcon.js
@@ -1,53 +1,11 @@
+import classList from '../utils/get-class-list-from-props'
 import convert from '../converter'
 import { icon, parse } from '@fortawesome/fontawesome-svg-core'
 import log from '../logger'
+import normalizeIconArgs from '../utils/normalize-icon-args'
+import objectWithKey from '../utils/object-with-key'
 import PropTypes from 'prop-types'
 import React from 'react'
-
-function objectWithKey(key, value) {
-  return (Array.isArray(value) && value.length > 0) ||
-    (!Array.isArray(value) && value)
-    ? { [key]: value }
-    : {}
-}
-
-function classList(props) {
-  let classes = {
-    'fa-spin': props.spin,
-    'fa-pulse': props.pulse,
-    'fa-fw': props.fixedWidth,
-    'fa-inverse': props.inverse,
-    'fa-border': props.border,
-    'fa-li': props.listItem,
-    'fa-flip-horizontal': props.flip === 'horizontal' || props.flip === 'both',
-    'fa-flip-vertical': props.flip === 'vertical' || props.flip === 'both',
-    [`fa-${props.size}`]: props.size !== null,
-    [`fa-rotate-${props.rotation}`]: props.rotation !== null,
-    [`fa-pull-${props.pull}`]: props.pull !== null
-  }
-
-  return Object.keys(classes)
-    .map(key => (classes[key] ? key : null))
-    .filter(key => key)
-}
-
-function normalizeIconArgs(icon) {
-  if (icon === null) {
-    return null
-  }
-
-  if (typeof icon === 'object' && icon.prefix && icon.iconName) {
-    return icon
-  }
-
-  if (Array.isArray(icon) && icon.length === 2) {
-    return { prefix: icon[0], iconName: icon[1] }
-  }
-
-  if (typeof icon === 'string') {
-    return { prefix: 'fas', iconName: icon }
-  }
-}
 
 export default function FontAwesomeIcon({ forwardedRef, ...props }) {
   const { icon: iconArgs, mask: maskArgs, symbol, className, title } = props
@@ -82,6 +40,7 @@ export default function FontAwesomeIcon({ forwardedRef, ...props }) {
   const extraProps = { ref: forwardedRef }
 
   Object.keys(props).forEach(key => {
+    // eslint-disable-next-line no-prototype-builtins
     if (!FontAwesomeIcon.defaultProps.hasOwnProperty(key)) {
       extraProps[key] = props[key]
     }
@@ -145,7 +104,9 @@ FontAwesomeIcon.propTypes = {
 
   title: PropTypes.string,
 
-  transform: PropTypes.oneOfType([PropTypes.string, PropTypes.object])
+  transform: PropTypes.oneOfType([PropTypes.string, PropTypes.object]),
+
+  swapOpacity: PropTypes.bool
 }
 
 FontAwesomeIcon.defaultProps = {
@@ -164,7 +125,8 @@ FontAwesomeIcon.defaultProps = {
   spin: false,
   symbol: false,
   title: '',
-  transform: null
+  transform: null,
+  swapOpacity: false
 }
 
 const convertCurry = convert.bind(null, React.createElement)